requirements.txt
click~=8.1.3
dacite~=1.8.0
Pillow~=10.1.0
pyclip~=0.7.0
pywin32~=303
requests~=2.26.0
setuptools~=63.3.0
torch~=1.13.1+cu116
torchvision~=0.14.1+cu116
